include: move RaucSlot definition etc. to slot.h
